requirements.txt
# ============================================================
# ASCSPipeline — Python Dependencies
# Target: Streamlit Community Cloud (Debian Linux, Python 3.11+)
# ============================================================

# --- Core Framework ---
streamlit>=1.35.0

# --- Smart Contract Static Analysis ---
# Slither: Trail of Bits static analyzer for Solidity
slither-analyzer>=0.10.0

# --- Solidity Compiler Management ---
# py-solc-x: Downloads and manages solc versions without system apt
py-solc-x>=2.0.0

# --- AI Inference ---
# Groq: High-speed LLM inference (Llama, Mixtral)
groq>=0.9.0

# --- RAG / Retrieval ---
# scikit-learn: TF-IDF vectorizer + cosine similarity for the knowledge base.
# Replaces sentence-transformers + faiss-cpu + torch (~1.5 GB) with a
# lightweight alternative that is perfectly adequate for 16 knowledge chunks.
scikit-learn>=1.4.0

# --- HTTP / Binary Download ---
# Requests: Used by setupTools.py to download the Aderyn binary
requests>=2.31.0

# --- Numerics ---
numpy>=1.26.0
